Await finishing using a promise instead of polling state
Fixes #209 and #298

Previously, we were using the sleepUntil utility to await the end state of the sync before quitting. sleepUntil loops internally with a sleep(0), which uses a lot of CPU! This updates us to use a promise that we manually resolve instead, which means the node process can actually sleep while nothing is happening.